Sun / Moon now working, cleaned up shaders, simulate time
Made sun and moon work based on the time of day, rotates around
the world (when active). Removed excess vertex shaders which were
just duplicates of the original Cube vertex shader. Made sun and
moon reference the billboard shader. Time now dynamic based on the
speedup and if toggled or not.